Scrapfly acquisition revisions management
- use base class for acquisition management commands
- move common code
- do not create revision if not changes made
- manage potential error on geocode
